Smart_PPT_Factory/requirements.txt
python-pptx>=0.6.21
google-genai>=0.2.0
pypdf>=3.0.0
python-dotenv>=1.0.0
